Permit Devise.allow_unconfirmed_access to be nil
Devise currently permits indefinite unconfirmed access when
`Devise.allow_unconfirmed_access` is set to `nil`.

See:
https://github.com/plataformatec/devise/blob/14863ba4c92cd9781a961be0486f0ea7dfe84144/lib/generators/templates/devise.rb#L139-L140

However, at present setting this parameter to nil results in NoMethodErrors
when evaluating the following:

```
def email=(new_email)
  multi_email.change_primary_email_to(new_email, allow_unconfirmed: Devise.allow_unconfirmed_access_for > 0.days)
end
```

This commit creates a new method that checks to ensure if the Devise
configuration permits any period in which an unconfirmed may log in
and uses it to determine whether or not an unconfirmed email address
may be added to a user.